Implement method that maps Category to CategoryEntity
