+++ backend/spring-app/spring-app/src/main/java/com/avdo/spring/app/entity/CategoryEntity.java
@@ -29,7 +29,7 @@ public class CategoryEntity {
 
     public Category toDomainModel() {
         List<Product> products = (this.productEntities == null ? Collections.emptyList() : this.productEntities.stream()
-                .map(Product::toDomainModel)
+                .map(ProductEntity::toDomainModel)
                 .collect(Collectors.toList()));
         return Category.builder()
                 .id(this.id)
@@ -39,10 +39,14 @@ public class CategoryEntity {
     }
 
     public static CategoryEntity fromCategory(Category category) {
+        List<ProductEntity> productEntities = (category.getProducts() == null ? Collections.emptyList() : category.getProducts().stream()
+                .map(ProductEntity::fromProduct) // You need to implement this method in ProductEntity
+                .collect(Collectors.toList()));
+
         CategoryEntity categoryEntity = new CategoryEntity();
         categoryEntity.setId(category.getId());
         categoryEntity.setName(category.getName());
-        categoryEntity.setProductEntities(category.getProducts());
+        categoryEntity.setProductEntities(productEntities);
         return categoryEntity;
     }
